feat(wasm): implement index and base-table queries
Extract query_items, query_gsi_items, and query_lsi_items into shared builders
(key condition, composite pagination cursors, ordering, limit) and implement
them on the wasm backend. Queries now run on both backends from one source of
SQL; scans follow.
